+++ src/main/java/com/networknt/rule/IAction.java
@@ -16,13 +16,7 @@ public interface IAction {
      * @param actionValues action values
      * @throws RuleEngineException exception thrown during performAction execution.
      */
-    default void performAction(String ruleId, String actionId, Map<String, Object> objMap, Map<String, Object> resultMap, Collection<RuleActionValue> actionValues) throws RuleEngineException {
-        throw new UnsupportedOperationException("Action must implement collection or map actionValues");
-    }
-
-    default void performAction(String ruleId, String actionId, Map<String, Object> objMap, Map<String, Object> resultMap, Map<String, Object> actionValues) throws RuleEngineException {
-        performAction(ruleId, actionId, objMap, resultMap, RuleAction.toActionValueCollection(actionValues));
-    }
+    void performAction(String ruleId, String actionId, Map<String, Object> objMap, Map<String, Object> resultMap, Collection<RuleActionValue> actionValues) throws RuleEngineException;
 
     /**
      * post action to be performed
@@ -35,9 +29,5 @@ public interface IAction {
      */
     default void postPerformAction(String ruleId, String actionId, Map<String, Object> objMap, Map<String, Object> resultMap, Collection<RuleActionValue> actionValues) throws RuleEngineException {
         // NOOP for classes implement IAction.
-    };
-
-    default void postPerformAction(String ruleId, String actionId, Map<String, Object> objMap, Map<String, Object> resultMap, Map<String, Object> actionValues) throws RuleEngineException {
-        postPerformAction(ruleId, actionId, objMap, resultMap, RuleAction.toActionValueCollection(actionValues));
     };
 }